Fix style issues
