+++ lib/mcp_lite/mcp_lite.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+begin
+  require_relative "mcp_lite.so"
+rescue LoadError
+  # The native extension is not built in the test environment.
+end
